Remove add button as new row is always present
Not sure about the impact of this decision. What about if the list
has millons of entries? Here's the problem. If add a new entry
then I want to see it after creating it. But, where?
At the beginning? At the end of the table? What about if data is sorted?